feat: shift business logic to module
