+++ BadgeMagicModule/src/commonMain/kotlin/org/fossasia/badgemagic/data/BadgeConfig.kt
@@ -1,9 +1,7 @@
-package org.fossasia.badgemagic.data.fragments
+package org.fossasia.badgemagic.data
 
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-import org.fossasia.badgemagic.data.device.model.Mode
-import org.fossasia.badgemagic.data.device.model.Speed
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 
 const val CONF_HEX_STRINGS = "hex_strings"
 const val CONF_SPEED = "speed"
@@ -12,23 +10,23 @@ const val CONF_FLASH = "flash"
 const val CONF_MARQUEE = "marquee"
 const val CONF_INVERTED = "inverted"
 
-@JsonClass(generateAdapter = true)
+@Serializable
 class BadgeConfig {
-    @Json(name = CONF_HEX_STRINGS)
+    @SerialName(CONF_HEX_STRINGS)
     var hexStrings: List<String> = mutableListOf()
 
-    @Json(name = CONF_SPEED)
+    @SerialName(CONF_SPEED)
     var speed: Speed = Speed.ONE
 
-    @Json(name = CONF_MODE)
+    @SerialName(CONF_MODE)
     var mode: Mode = Mode.LEFT
 
-    @Json(name = CONF_FLASH)
+    @SerialName(CONF_FLASH)
     var isFlash: Boolean = false
 
-    @Json(name = CONF_MARQUEE)
+    @SerialName(CONF_MARQUEE)
     var isMarquee: Boolean = false
 
-    @Json(name = CONF_INVERTED)
+    @SerialName(CONF_INVERTED)
     var isInverted: Boolean = false
 }
