+++ BadgeMagicModule/src/commonMain/kotlin/org/fossasia/badgemagic/data/ConfigInfo.kt
@@ -1,3 +1,3 @@
-package org.fossasia.badgemagic.data.fragments
+package org.fossasia.badgemagic.data
 
 data class ConfigInfo(val badgeJSON: String, val fileName: String)
