+++ BadgeMagicModule/src/commonMain/kotlin/org/fossasia/badgemagic/data/DataToSend.kt
@@ -1,0 +1,6 @@
+package org.fossasia.badgemagic.data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DataToSend(val messages: List<Message>)
