+++ BadgeMagicModule/src/commonMain/kotlin/org/fossasia/badgemagic/data/Message.kt
@@ -1,3 +1,6 @@
-package org.fossasia.badgemagic.data.device.model
+package org.fossasia.badgemagic.data
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class Message(val hexStrings: List<String>, val flash: Boolean = false, val marquee: Boolean = false, val speed: Speed = Speed.ONE, val mode: Mode = Mode.LEFT)
