+++ app/src/main/java/org/fossasia/badgemagic/core/bluetooth/Constants.kt
@@ -1,4 +1,4 @@
-package org.fossasia.badgemagic.data.device.bluetooth
+package org.fossasia.badgemagic.core.bluetooth
 
 import java.util.UUID
