+++ app/src/main/java/org/fossasia/badgemagic/data/ModeInfo.kt
@@ -1,6 +1,5 @@
 package org.fossasia.badgemagic.data
 
 import androidx.annotation.DrawableRes
-import org.fossasia.badgemagic.data.device.model.Mode
 
 data class ModeInfo(@DrawableRes val drawableID: Int, val mode: Mode)
