--- app/src/main/java/org/fossasia/badgemagic/data/device/model/DataToSend.kt
@@ -1,3 +1,0 @@
-package org.fossasia.badgemagic.data.device.model
-
-data class DataToSend(val messages: List<Message>)
