+++ app/src/main/java/org/fossasia/badgemagic/ui/base/BaseFragment.kt
@@ -1,7 +1,7 @@
 package org.fossasia.badgemagic.ui.base
 
 import androidx.fragment.app.Fragment
-import org.fossasia.badgemagic.data.device.model.DataToSend
+import org.fossasia.badgemagic.data.DataToSend
 
 open class BaseFragment : Fragment(), BaseFragmentInterface {
     override fun initializePreview() {
