+++ build.gradle
@@ -13,6 +13,8 @@ buildscript {
         classpath "com.android.tools.build:gradle:3.5.3"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$versions.kotlin_version"
         classpath "gradle.plugin.com.github.b3er.local.properties:local-properties-plugin:1.1"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$versions.kotlin_version"
+        classpath "org.jetbrains.kotlin:kotlin-serialization:$versions.kotlin_version"
     }
 }
