+++ versions.gradle
@@ -5,7 +5,7 @@ versions.buildTools = '29.0.2'
 versions.minSdk = 21
 versions.targetSdk = 29
 
-versions.kotlin_version = '1.3.41'
+versions.kotlin_version = '1.3.61'
 
 versions.koin_version = '2.0.1'
 versions.moshi_version = '1.8.0'
